+++ djstripe/migrations/0002_auto_20180627_1121.py
@@ -456,4 +456,266 @@ class Migration(migrations.Migration):
             old_name="stripe_id",
             new_name="id",
         ),
+
+        # Update all text-type fields to non-null CharField blank=True default=""
+        migrations.AlterField(
+            model_name='account',
+            name='business_name',
+            field=models.CharField(blank=True, default='', help_text='The publicly visible name of the business', max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='account',
+            name='business_primary_color',
+            field=models.CharField(blank=True, default='', help_text='A CSS hex color value representing the primary branding color for this account', max_length=7),
+        ),
+        migrations.AlterField(
+            model_name='account',
+            name='business_url',
+            field=models.CharField(blank=True, default='', help_text='The publicly visible website of the business', max_length=200),
+        ),
+        migrations.AlterField(
+            model_name='account',
+            name='payout_statement_descriptor',
+            field=models.CharField(blank=True, default='', help_text='The text that appears on the bank account statement for payouts.', max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='account',
+            name='product_description',
+            field=models.CharField(blank=True, default='', help_text='Internal-only description of the product sold or service provided by the business. It’s used by Stripe for risk and underwriting purposes.', max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='account',
+            name='support_url',
+            field=models.CharField(blank=True, default='', help_text='A publicly shareable URL that provides support for this account', max_length=200),
+        ),
+        migrations.AlterField(
+            model_name='bankaccount',
+            name='account_holder_name',
+            field=models.CharField(blank=True, default='', help_text='The name of the person or business that owns the bank account.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_city',
+            field=models.TextField(blank=True, default='', help_text='City/District/Suburb/Town/Village.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_country',
+            field=models.CharField(blank=True, default='', help_text='Billing address country.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_line1',
+            field=models.CharField(blank=True, default='', help_text='Street address/PO Box/Company name.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_line1_check',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.CardCheckResult, help_text='If `address_line1` was provided, results of the check.', max_length=11),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_line2',
+            field=models.CharField(blank=True, default='', help_text='Apartment/Suite/Unit/Building.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_state',
+            field=models.CharField(blank=True, default='', help_text='State/County/Province/Region.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_zip',
+            field=models.CharField(blank=True, default='', help_text='ZIP or postal code.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='address_zip_check',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.CardCheckResult, help_text='If `address_zip` was provided, results of the check.', max_length=11),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='country',
+            field=models.CharField(blank=True, default='', help_text='Two-letter ISO code representing the country of the card.', max_length=2),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='cvc_check',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.CardCheckResult, help_text='If a CVC was provided, results of the check.', max_length=11),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='dynamic_last4',
+            field=models.CharField(blank=True, default='', help_text='(For tokenized numbers only.) The last four digits of the device account number.', max_length=4),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='fingerprint',
+            field=models.CharField(blank=True, default='', help_text='Uniquely identifies this particular card number.', max_length=16),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='name',
+            field=models.CharField(blank=True, default='', help_text='Cardholder name.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='card',
+            name='tokenization_method',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.CardTokenizationMethod, help_text='If the card number is tokenized, this is the method that was used.', max_length=11),
+        ),
+        migrations.AlterField(
+            model_name='charge',
+            name='failure_code',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.ApiErrorCode, help_text='Error code explaining reason for charge failure if available.', max_length=20),
+        ),
+        migrations.AlterField(
+            model_name='charge',
+            name='failure_message',
+            field=models.CharField(blank=True, default='', help_text='Message to user further explaining reason for charge failure if available.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='charge',
+            name='receipt_email',
+            field=models.CharField(blank=True, default='', help_text='The email address that the receipt for this charge was sent to.', max_length=800),
+        ),
+        migrations.AlterField(
+            model_name='charge',
+            name='receipt_number',
+            field=models.CharField(blank=True, default='', help_text='The transaction number that appears on email receipts sent for this charge.', max_length=14),
+        ),
+        migrations.AlterField(
+            model_name='charge',
+            name='statement_descriptor',
+            field=models.CharField(blank=True, default='', help_text='An arbitrary string to be displayed on your customer\'s credit card statement. The statement description may not include <>"\' characters, and will appear on your customer\'s statement in capital letters. Non-ASCII characters are automatically stripped. While most banks display this information consistently, some may display it incorrectly or not at all.', max_length=22),
+        ),
+        migrations.AlterField(
+            model_name='charge',
+            name='transfer_group',
+            field=models.CharField(blank=True, default='', help_text='A string that identifies this transaction as part of a group.', max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='coupon',
+            name='name',
+            field=models.CharField(blank=True, default='', help_text='Name of the coupon displayed to customers on for instance invoices or receipts.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='customer',
+            name='business_vat_id',
+            field=models.CharField(blank=True, default='', help_text="The customer's VAT identification number.", max_length=20),
+        ),
+        migrations.AlterField(
+            model_name='customer',
+            name='currency',
+            field=djstripe.fields.StripeCurrencyCodeField(default='', help_text='The currency the customer can be charged in for recurring billing purposes', max_length=3),
+        ),
+        migrations.AlterField(
+            model_name='customer',
+            name='email',
+            field=models.CharField(blank=True, db_index=True, default='', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='event',
+            name='idempotency_key',
+            field=models.TextField(blank=True, default=''),
+        ),
+        migrations.AlterField(
+            model_name='event',
+            name='request_id',
+            field=models.CharField(blank=True, default='', help_text="Information about the request that triggered this event, for traceability purposes. If empty string then this is an old entry without that data. If Null then this is not an old entry, but a Stripe 'automated' event with no associated request.", max_length=50),
+        ),
+        migrations.AlterField(
+            model_name='invoice',
+            name='hosted_invoice_url',
+            field=models.CharField(blank=True, default='', help_text='The URL for the hosted invoice page, which allows customers to view and pay an invoice. If the invoice has not been frozen yet, this will be null.', max_length=799),
+        ),
+        migrations.AlterField(
+            model_name='invoice',
+            name='invoice_pdf',
+            field=models.CharField(blank=True, default='', help_text='The link to download the PDF for the invoice. If the invoice has not been frozen yet, this will be null.', max_length=799),
+        ),
+        migrations.AlterField(
+            model_name='invoice',
+            name='number',
+            field=models.CharField(blank=True, default='', help_text='A unique, identifying string that appears on emails sent to the customer for this invoice. This starts with the customer’s unique invoice_prefix if it is specified.', max_length=64),
+        ),
+        migrations.AlterField(
+            model_name='invoice',
+            name='statement_descriptor',
+            field=models.CharField(blank=True, default='', help_text='An arbitrary string to be displayed on your customer\'s credit card statement. The statement description may not include <>"\' characters, and will appear on your customer\'s statement in capital letters. Non-ASCII characters are automatically stripped. While most banks display this information consistently, some may display it incorrectly or not at all.', max_length=22),
+        ),
+        migrations.AlterField(
+            model_name='payout',
+            name='failure_code',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.PayoutFailureCode, help_text='Error code explaining reason for transfer failure if available. See https://stripe.com/docs/api/python#transfer_failures.', max_length=23),
+        ),
+        migrations.AlterField(
+            model_name='payout',
+            name='failure_message',
+            field=models.TextField(blank=True, default='', help_text='Message to user further explaining reason for payout failure if available.'),
+        ),
+        migrations.AlterField(
+            model_name='payout',
+            name='statement_descriptor',
+            field=models.CharField(blank=True, default='', help_text="Extra information about a payout to be displayed on the user's bank statement.", max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='plan',
+            name='aggregate_usage',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.PlanAggregateUsage, help_text='Specifies a usage aggregation strategy for plans of usage_type=metered. Allowed values are `sum` for summing up all usage during a period, `last_during_period` for picking the last usage record reported within a period, `last_ever` for picking the last usage record ever (across period bounds) or max which picks the usage record with the maximum reported usage during a period. Defaults to `sum`.', max_length=18),
+        ),
+        migrations.AlterField(
+            model_name='plan',
+            name='billing_scheme',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.PlanBillingScheme, help_text='Describes how to compute the price per period. Either `per_unit` or `tiered`. `per_unit` indicates that the fixed amount (specified in amount) will be charged per unit in quantity (for plans with `usage_type=licensed`), or per unit of total usage (for plans with `usage_type=metered`). `tiered` indicates that the unit pricing will be computed using a tiering strategy as defined using the tiers and tiers_mode attributes.', max_length=8),
+        ),
+        migrations.AlterField(
+            model_name='plan',
+            name='nickname',
+            field=models.CharField(blank=True, default='', help_text='A brief description of the plan, hidden from customers.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='product',
+            name='caption',
+            field=models.CharField(blank=True, default='', help_text='A short one-line description of the product, meant to be displayableto the customer. Only applicable to products of `type=good`.', max_length=5000),
+        ),
+        migrations.AlterField(
+            model_name='product',
+            name='statement_descriptor',
+            field=models.CharField(blank=True, default='', help_text="Extra information about a product which will appear on your customer's credit card statement. In the case that multiple products are billed at once, the first statement descriptor will be used. Only available on products of type=`service`.", max_length=22),
+        ),
+        migrations.AlterField(
+            model_name='product',
+            name='unit_label',
+            field=models.CharField(blank=True, default='', max_length=12),
+        ),
+        migrations.AlterField(
+            model_name='refund',
+            name='failure_reason',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.RefundFailureReason, help_text='If the refund failed, the reason for refund failure if known.', max_length=24),
+        ),
+        migrations.AlterField(
+            model_name='refund',
+            name='reason',
+            field=djstripe.fields.StripeEnumField(blank=True, default='', enum=djstripe.enums.RefundReason, help_text='Reason for the refund.', max_length=21),
+        ),
+        migrations.AlterField(
+            model_name='refund',
+            name='receipt_number',
+            field=models.CharField(blank=True, default='', help_text='The transaction number that appears on email receipts sent for this charge.', max_length=9),
+        ),
+        migrations.AlterField(
+            model_name='source',
+            name='currency',
+            field=djstripe.fields.StripeCurrencyCodeField(blank=True, default='', help_text='Three-letter ISO currency code', max_length=3),
+        ),
+        migrations.AlterField(
+            model_name='source',
+            name='statement_descriptor',
+            field=models.CharField(blank=True, default='', help_text="Extra information about a source. This will appear on your customer's statement every time you charge the source.", max_length=255),
+        ),
+        migrations.AlterField(
+            model_name='transfer',
+            name='transfer_group',
+            field=models.CharField(blank=True, default='', help_text='A string that identifies this transaction as part of a group.', max_length=255),
+        ),
     ]